+++ src/main/java/kahlua/KahluaProject/controller/ReservationController.java
@@ -52,20 +52,4 @@ public class ReservationController {
 
         return ApiResponse.onSuccess(reservationService.getByDate(date));
     }
-
-    @GetMapping("/v1/reservation/check")
-    @Operation(summary = "사용자 예약내역 조회", description = "사용자의 전체 예약내역을 예약 날짜 오름차순으로 조회합니다." +
-            "<br> 같은 날짜에 대해서는 시작 시간 오름차순으로 정렬합니다.")
-    public ApiResponse<ReservationListResponse> getReservationList(@AuthenticationPrincipal AuthDetails authDetails) {
-
-        return ApiResponse.onSuccess(reservationService.getByUser(authDetails.user()));
-    }
-
-    @DeleteMapping("/v1/reservation/check/{reservationId}")
-    @Operation(summary = "예약내역 삭제", description = "지정한 예약내역을 삭제합니다.")
-    public ApiResponse<?> delete(@PathVariable Long reservationId) {
-
-        reservationService.delete(reservationId);
-        return ApiResponse.onSuccess("예약내역 삭제 성공");
-    }
 }
